Refactor: make ChatMessageListItem top-level class
