+++ apps/desktop/desktop/src/main/java/bisq/desktop/main/content/components/ChatMessagesComponent.java
@@ -40,6 +40,7 @@ import bisq.desktop.components.controls.BisqTextArea;
 import bisq.desktop.components.controls.BisqTooltip;
 import bisq.desktop.components.overlay.Popup;
 import bisq.desktop.main.content.chat.ChatUtil;
+import bisq.desktop.main.content.components.chatMessages.ChatMessageListItem;
 import bisq.i18n.Res;
 import bisq.offer.Direction;
 import bisq.offer.bisq_easy.BisqEasyOffer;
@@ -98,7 +99,7 @@ public class ChatMessagesComponent {
         controller.mentionUserHandler(userProfile);
     }
 
-    public void setSearchPredicate(Predicate<? super ChatMessagesListView.ChatMessageListItem<? extends ChatMessage>> predicate) {
+    public void setSearchPredicate(Predicate<? super ChatMessageListItem<? extends ChatMessage>> predicate) {
         controller.chatMessagesListView.setSearchPredicate(predicate);
     }
